+++ js/math.js
@@ -1,0 +1,7 @@
+export function getRandomNum(min, max) {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+export function getRandomElement(arr) {
+  return arr[Math.floor(Math.random() * arr.length)];
+}
